requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cyberpi==0.0.5
cycler==0.10.0
discord.py==1.7.3
Django==3.2.8
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
fastapi==0.70.0
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
kiwisolver==1.3.2
makeblock==0.1.7
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
multidict==5.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.26
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.21.2
orjson==3.6.4
packaging==21.2
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.22
protobuf==3.18.1
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.8.2
pygame==2.0.3
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
toml==0.10.2
tornado==6.1
touch==2020.12.3
traitlets==5.1.1
typing_extensions==4.0.0
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
values==2020.12.3
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
XlsxWriter==3.0.1
yarl==1.7.2
